refac: Segregate application service model stuff (#743)
