+++ source/dotnet/Packages/Contracts/Contracts.csproj
@@ -36,7 +36,7 @@ limitations under the License.
 
   <PropertyGroup>
       <PackageId>Energinet.DataHub.Wholesale.Contracts</PackageId>
-      <PackageVersion>1.0.3$(VersionSuffix)</PackageVersion>
+      <PackageVersion>2.0.0$(VersionSuffix)</PackageVersion>
       <Title>Wholesale Contracts</Title>
       <Company>Energinet-DataHub</Company>
       <Authors>Energinet-DataHub</Authors>
