+++ source/dotnet/Services/Application/Application.csproj
@@ -34,31 +34,31 @@ limitations under the License.
 
   <ItemGroup>
     <Compile Include="..\..\Contracts\BatchRequestDto.cs">
-      <Link>Batches\BatchRequestDto.cs</Link>
+      <Link>Batches\Model\BatchRequestDto.cs</Link>
     </Compile>
     <Compile Include="..\..\Contracts\BatchSearchDto.cs">
-      <Link>Batches\BatchSearchDto.cs</Link>
-    </Compile>
-    <Compile Include="..\..\Contracts\ProcessStepMeteringPointType.cs">
-      <Link>ProcessResult\ProcessStepMeteringPointType.cs</Link>
-    </Compile>
-    <Compile Include="..\..\Contracts\ProcessStepResultDto.cs">
-      <Link>ProcessResult\ProcessStepResultDto.cs</Link>
-    </Compile>
-    <Compile Include="..\..\Contracts\ProcessStepResultRequestDto.cs">
-      <Link>ProcessResult\ProcessStepResultRequestDto.cs</Link>
-    </Compile>
-    <Compile Include="..\..\Contracts\ProcessStepType.cs">
-      <Link>ProcessResult\ProcessStepType.cs</Link>
-    </Compile>
-    <Compile Include="..\..\Contracts\ProcessType.cs">
-      <Link>Processes\ProcessType.cs</Link>
+      <Link>Batches\Model\BatchSearchDto.cs</Link>
     </Compile>
     <Compile Include="..\..\Contracts\BatchState.cs">
-      <Link>Batches\BatchState.cs</Link>
+      <Link>Batches\Model\BatchState.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Contracts\ProcessStepMeteringPointType.cs">
+      <Link>ProcessResult\Model\ProcessStepMeteringPointType.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Contracts\ProcessStepResultDto.cs">
+      <Link>ProcessResult\Model\ProcessStepResultDto.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Contracts\ProcessStepResultRequestDto.cs">
+      <Link>ProcessResult\Model\ProcessStepResultRequestDto.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Contracts\ProcessStepType.cs">
+      <Link>ProcessResult\Model\ProcessStepType.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Contracts\ProcessType.cs">
+      <Link>Processes\Model\ProcessType.cs</Link>
     </Compile>
     <Compile Include="..\..\Contracts\TimeSeriesPointDto.cs">
-      <Link>ProcessResult\TimeSeriesPointDto.cs</Link>
+      <Link>ProcessResult\Model\TimeSeriesPointDto.cs</Link>
     </Compile>
   </ItemGroup>
