+++ source/dotnet/Services/Application/Batches/IBatchCompletedPublisher.cs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Application.Batches;
+using Energinet.DataHub.Wholesale.Application.Batches.Model;
 
-namespace Energinet.DataHub.Wholesale.Application.Infrastructure;
+namespace Energinet.DataHub.Wholesale.Application.Batches;
 
 public interface IBatchCompletedPublisher
 {
