+++ source/dotnet/Services/Application/Batches/Model/BatchDto.cs
@@ -12,11 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
+using Energinet.DataHub.Wholesale.Contracts;
 
-namespace Energinet.DataHub.Wholesale.Application.Batches;
+namespace Energinet.DataHub.Wholesale.Application.Batches.Model;
 
-public interface IBatchDtoMapper
-{
-    BatchDto Map(Batch batch);
-}
+/// <summary>
+/// An immutable batch.
+/// </summary>
+public sealed record BatchDto(
+    long? RunId,
+    Guid BatchId,
+    DateTimeOffset PeriodStart,
+    DateTimeOffset PeriodEnd,
+    DateTimeOffset? ExecutionTimeStart,
+    DateTimeOffset? ExecutionTimeEnd,
+    BatchState ExecutionState,
+    bool AreSettlementReportsCreated,
+    string[] GridAreaCodes);
