+++ source/dotnet/Services/Application/Batches/Model/BatchDtoMapper.cs
@@ -16,7 +16,7 @@ using Energinet.DataHub.Wholesale.Contracts;
 using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
 using Energinet.DataHub.Wholesale.Domain.GridAreaAggregate;
 
-namespace Energinet.DataHub.Wholesale.Application.Batches;
+namespace Energinet.DataHub.Wholesale.Application.Batches.Model;
 
 public class BatchDtoMapper : IBatchDtoMapper
 {
