+++ source/dotnet/Services/Application/ProcessResult/Model/ProcessStepResultMapper.cs
@@ -15,7 +15,7 @@
 using Energinet.DataHub.Wholesale.Contracts;
 using Energinet.DataHub.Wholesale.Domain.ProcessStepResultAggregate;
 
-namespace Energinet.DataHub.Wholesale.Application.ProcessResult;
+namespace Energinet.DataHub.Wholesale.Application.ProcessResult.Model;
 
 public class ProcessStepResultMapper : IProcessStepResultMapper
 {
