+++ source/dotnet/Services/Application/Processes/IProcessCompletedIntegrationEventPublisher.cs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Application.Processes;
+using Energinet.DataHub.Wholesale.Application.Processes.Model;
 
-namespace Energinet.DataHub.Wholesale.Application.Infrastructure;
+namespace Energinet.DataHub.Wholesale.Application.Processes;
 
 public interface IProcessCompletedIntegrationEventPublisher
 {
