+++ source/dotnet/Services/Application/Processes/IProcessCompletedPublisher.cs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.Wholesale.Application.Processes.Model;
+
 namespace Energinet.DataHub.Wholesale.Application.Processes;
 
 public interface IProcessCompletedPublisher
