+++ source/dotnet/Services/Application/Processes/Model/ProcessCompletedEventDtoFactory.cs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Application.Batches;
+using Energinet.DataHub.Wholesale.Application.Batches.Model;
 
-namespace Energinet.DataHub.Wholesale.Application.Processes;
+namespace Energinet.DataHub.Wholesale.Application.Processes.Model;
 
 public class ProcessCompletedEventDtoFactory : IProcessCompletedEventDtoFactory
 {
