+++ source/dotnet/Services/Application/Processes/Model/ProcessTypeMapper.cs
@@ -12,14 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Domain.ProcessAggregate;
-using NodaTime;
+using Energinet.DataHub.Wholesale.Contracts;
 
-namespace Energinet.DataHub.Wholesale.Application.Batches;
+namespace Energinet.DataHub.Wholesale.Application.Processes.Model;
 
-public sealed record BatchCompletedEventDto(
-    Guid BatchId,
-    List<string> GridAreaCodes,
-    ProcessType ProcessType,
-    Instant PeriodStart,
-    Instant PeriodEnd);
+public class ProcessTypeMapper : IProcessTypeMapper
+{
+    public ProcessType MapFrom(Domain.ProcessAggregate.ProcessType processType)
+    {
+        return processType switch
+        {
+            Domain.ProcessAggregate.ProcessType.BalanceFixing => ProcessType.BalanceFixing,
+            _ => throw new NotImplementedException($"Cannot map process type '{processType}"),
+        };
+    }
+}
