+++ source/dotnet/Services/Application/SettlementReport/Model/SettlementReportDto.cs
@@ -12,6 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Application.SettlementReport;
+namespace Energinet.DataHub.Wholesale.Application.SettlementReport.Model;
 
 public record SettlementReportDto(Stream Stream);
