+++ source/dotnet/Services/ProcessManager/Endpoints/CreateSettlementReportsEndpoint.cs
@@ -14,6 +14,7 @@
 
 using Energinet.DataHub.Core.JsonSerialization;
 using Energinet.DataHub.Wholesale.Application.Batches;
+using Energinet.DataHub.Wholesale.Application.Batches.Model;
 using Energinet.DataHub.Wholesale.Application.SettlementReport;
 using Microsoft.Azure.Functions.Worker;
