+++ source/dotnet/Services/ProcessManager/Endpoints/PublishProcessCompletedEventsEndpoint.cs
@@ -14,6 +14,7 @@
 
 using Energinet.DataHub.Core.JsonSerialization;
 using Energinet.DataHub.Wholesale.Application.Batches;
+using Energinet.DataHub.Wholesale.Application.Batches.Model;
 using Energinet.DataHub.Wholesale.Application.Processes;
 using Microsoft.Azure.Functions.Worker;
