+++ source/dotnet/Services/ProcessManager/Endpoints/PublishProcessCompletedIntegrationEventEndpoint.cs
@@ -14,6 +14,7 @@
 
 using Energinet.DataHub.Core.JsonSerialization;
 using Energinet.DataHub.Wholesale.Application.Processes;
+using Energinet.DataHub.Wholesale.Application.Processes.Model;
 using Microsoft.Azure.Functions.Worker;
 
 namespace Energinet.DataHub.Wholesale.ProcessManager.Endpoints;
